impl(internal/sidekick): move tests
Move all tests from sidekick to internal/sidekick other than TestMain.
